openai-client/src/test/java/com/lipcha/model/completion/CompletionTest.java
package com.lipcha.model.completion;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.lipcha.util.ParseUtils;
import com.lipcha.util.TestFileUtils;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class CompletionTest {

	@Test
	void test_if_completion_parsed_and_serialized() throws JsonProcessingException {
		String jsonString = TestFileUtils.testFile("/model/completion.json");
		Completion completion = ParseUtils.MAPPER.readValue(jsonString, Completion.class);
		assertThat(ParseUtils.MAPPER.readTree(jsonString))
				.isEqualTo(ParseUtils.MAPPER.readTree(ParseUtils.MAPPER.writeValueAsString(completion)));
	}

}